Ensure dtype stays the same in NoiseAugment

+++ gunpowder/nodes/noise_augment.py
@@ -51,4 +51,9 @@ class NoiseAugment(BatchFilter):
 
         assert raw.data.dtype == np.float32 or raw.data.dtype == np.float64, "Noise augmentation requires float types for the raw array (not " + str(raw.data.dtype) + "). Consider using Normalize before."
         assert raw.data.min() >= -1 and raw.data.max() <= 1, "Noise augmentation expects raw values in [-1,1] or [0,1]. Consider using Normalize before."
-        raw.data = skimage.util.random_noise(raw.data, mode=self.mode, seed=self.seed, clip=self.clip, **self.kwargs )
+        raw.data = skimage.util.random_noise(
+            raw.data,
+            mode=self.mode,
+            seed=self.seed,
+            clip=self.clip,
+            **self.kwargs).astype(raw.data.dtype)
